% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{aREA.enrich}
\alias{aREA.enrich}
\title{COMPUTE AREA ENRICHMENT BETWEEN ALL PAIRWISE COMBINATIONS OF VIPER PROTEINS AND EVENTS}
\usage{
aREA.enrich(events.mat, vipermat)
}
\arguments{
\item{events.mat}{: A BINARY 0/1 MATRIX WITH SAMPLES AS COLUMNS AND ROWS AS GENES/EVENTS}

\item{vipermat}{: A MATRIX OF INFERRED VIPER ACTIVITIES WITH SAMPLES AS COLUMNS AND ROWS AS PROTEINS}
}
\value{
A MATRIX OF NETWORK ENRICHMENT SCORES (NES) WITH ROWS AS EVENT/GENE NAMES AND COLUMNS AS VIPER PROTEIN NAMES
}
\description{
COMPUTE AREA ENRICHMENT BETWEEN ALL PAIRWISE COMBINATIONS OF VIPER PROTEINS AND EVENTS
}
