% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{aREA.regulon_enrich}
\alias{aREA.regulon_enrich}
\title{Compute aREA enrichment for the proteins in a given regulon, against vipermat scores supplied}
\usage{
aREA.regulon_enrich(regulon, vipermat)
}
\arguments{
\item{regulon}{ARACNE regulon object}

\item{vipermat}{A VIPER network of inferred activity scores with columns as patient samples, and rows as proteins}
}
\value{
A MATRIX OF NETWORK ENRICHMENT SCORES (NES) WITH ROWS AS EVENT/GENE NAMES AND COLUMNS AS VIPER PROTEIN NAMES
}
\description{
Compute aREA enrichment for the proteins in a given regulon, against vipermat scores supplied
}
