% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{areaEnrich}
\alias{areaEnrich}
\title{aREA.enrich Compute aREA enrichment between all pairwise combinations of VIPER proteins and gene-level events}
\usage{
areaEnrich(events.mat, vipermat, event.type)
}
\arguments{
\item{events.mat}{A Binary 0/1 matrix with columns as samples, and rows as proteins}

\item{vipermat}{A VIPER network of inferred activity scores with columns as samples, and rows as proteins}

\item{event.type}{Name of the event type for printing purposes}
}
\value{
A matrix of enrichment scores with rows as event/gene names and columns as VIPER protein names
}
\description{
aREA.enrich Compute aREA enrichment between all pairwise combinations of VIPER proteins and gene-level events
}
