% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{bestVSworst}
\alias{bestVSworst}
\title{Best vs Worst}
\usage{
bestVSworst(
  equivalent.clusters,
  search.results,
  survival.data,
  name.res,
  progression.free
)
}
\arguments{
\item{equivalent.clusters}{vector indicating k of equivalent clusters}

\item{search.results}{clustering object from clusterRange}

\item{survival.data}{dataframe with survival data for each sample}

\item{name.res}{list with length of sample names and column name for samples}

\item{progression.free}{T/F about whether to use progression free instead of overall survival}
}
\value{
k for the top solution
}
\description{
Take clustering solutions with equivalent reliability distributions to the best
one and calculate which result has the highest p value between the best and 
worst surviving clusters
}
\keyword{internal}
