% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{cdf.pval}
\alias{cdf.pval}
\title{Empirical Cumulative Distribution Helper function}
\usage{
cdf.pval(vals, na_value = NA)
}
\arguments{
\item{vals}{original list of p-values to convert}

\item{na_value}{value to substitute for NAs if present. Default is NA}
}
\value{
vector of adjusted p-values
}
\description{
This is a wrapper function for the ecdf function in order to get a vector of 
the resulting values
}
\keyword{internal}
