% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{cdfPval}
\alias{cdfPval}
\title{Empirical Cumulative Distribution Helper function}
\usage{
cdfPval(vals, na.value = NA)
}
\arguments{
\item{vals}{original list of p-values to convert}

\item{na.value}{value to substitute for NAs if present. Default is NA}
}
\value{
vector of adjusted p-values
}
\description{
This is a wrapper function for the ecdf function in order to get a vector of 
the resulting values
}
\keyword{internal}
