% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{checkList}
\alias{checkList}
\title{Check List of Assays}
\usage{
checkList(assaylist)
}
\arguments{
\item{assaylist}{list of assays (viper, cnv, mut and fusion)}
}
\description{
Check List of Assays
}
\keyword{internal}
