% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{checkMAE}
\alias{checkMAE}
\title{Check MultiAssayExperiment}
\usage{
checkMAE(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
updated/filtered MAE
}
\description{
Check MultiAssayExperiment
}
\keyword{internal}
