% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{checkPathways}
\alias{checkPathways}
\title{Check Pathways}
\usage{
checkPathways(pathways, x, type)
}
\arguments{
\item{pathways}{A named list of lists. Each named list represents 
interactions between proteins (keys) and their associated partners}

\item{x}{the MAE or Assaylist}

\item{type}{whether x is MAE or Assaylist}
}
\value{
nothing
}
\description{
Check Pathways
}
\keyword{internal}
