% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{checkSurvivalNames}
\alias{checkSurvivalNames}
\title{Check survival names}
\usage{
checkSurvivalNames(search.results, survival.data)
}
\arguments{
\item{search.results}{clustering object from clusterRange}

\item{survival.data}{dataframe with survival data for each sample}
}
\value{
list with length of sample names and column name for samples
}
\description{
Function to check names of samples in survival data 
Needed mainly because names are shorter in patient data
}
\keyword{internal}
