% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma_arguments.R
\name{cluster}
\alias{cluster}
\alias{moma_cluster}
\title{Cluster penalty}
\usage{
moma_cluster(..., w = NULL, ADMM = FALSE, acc = FALSE, eps = 1e-10,
  ..., lambda = 0, select_scheme = "g")
}
\arguments{
\item{...}{Forces users to specify all arguments by name.}

\item{w}{A symmetric square matrix. \code{w[i, j]} is the \eqn{w_{ij}} described above.}

\item{ADMM}{A Boolean value. Set to \code{TRUE} to use ADMM, set to \code{FALSE} to use AMA. Defaults to FALSE.}

\item{acc}{A Boolean value. Set to \code{TRUE} to use the accelerated version of the algorithm.
Currently we support accelerated AMA only.}

\item{eps}{A small numeric value. The precision used when solving the proximal operator.}

\item{lambda}{A vector containing penalty values}

\item{select_scheme}{A char being either "b" (nested BIC search) or "g" (grid search).

MoMA provides a flexible framework for regularized multivariate analysis
with several tuning parameters for different forms of regularization.
To assist the user in selecting these parameters (\code{alpha_u},
\code{alpha_v}, \code{lambda_u}, \code{lambda_v}), we provide
two selection modes: grid search ("g") and nested BIC search ("b").
Grid search means we solve the problem
for all combinations of parameter values provided by the user.

To explain nested BIC search, we need to look into how the algorithm runs.
To find an (approximate) solution to a penalized SVD (Singular Value Decomposition) problem is to solve two
penalized regression problems iteratively. Let's call them problem u and problem v, which give
improving estimates of the right singular vector, \emph{u}, and the left singular vector, \emph{v}, respectively.
For each regression problem, we can select the optimal parameters
based on BIC.

The nested BIC search is essentially two 2-D searches. We start from SVD solutions, and then find the optimal
parameters for problem u, given current estimate of \emph{v}. Using the result from previous step, update
current estimate of \emph{u}, and then do the same thing for problem v,
that is, to find the optimal parameters for problem v given current estimate of \emph{u}. Repeat
the above until convergence or the maximal number of iterations has been reached.

Users are welcome to refer to section 3.1: Selection of Regularization Parameters
in the paper cited below.}
}
\value{
A \code{moma_sparsity_type} object, which is a list containing the values of \code{w},
\code{ADMM}, \code{acc} and \code{eps}.
}
\description{
Use this function to set the penalty function to
\deqn{\lambda \sum w_{ij} | x_{i} - x_{j} |,}
where \eqn{\lambda} is set by the \code{lambda} argument below.
}
\references{
Chi, Eric C., and Kenneth Lange. "Splitting Methods for Convex Clustering."
Journal of Computational and Graphical Statistics 24.4 (2015): 994-1013. \doi{10.1080/10618600.2014.948181}.
}
