% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.functions.r
\name{cnvScoreStouffer}
\alias{cnvScoreStouffer}
\title{diggit.interactions: list indexed by MR/TF name in Entrez Space
	each points to a named vector of NES / z-scores associated with entrez IDs for each interacting event.}
\usage{
cnvScoreStouffer(mapping, diggit.interactions, cytoband = TRUE,
  from.p = FALSE, pos.nes.only = TRUE)
}
\description{
diggit.interactions: list indexed by MR/TF name in Entrez Space
	each points to a named vector of NES / z-scores associated with entrez IDs for each interacting event.
}
