% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{conditional.p}
\alias{conditional.p}
\title{Get the conditional p-value of a gene}
\usage{
conditional.p(gene.name, condition.on, x)
}
\arguments{
\item{gene.name}{Character}

\item{condition.on}{named Vector of scores for the distribution we are conditioning ON}

\item{x}{named Vector of scores for the dependent distribution}
}
\value{
a numeric p-value between 0 and 1
}
\description{
Get the conditional p-value of a gene
}
