% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.old.r
\name{conditionalModel}
\alias{conditionalModel}
\title{Implements the conditional Bayes model to combine VIPER scores with diggit 
and pathway scores}
\usage{
conditionalModel(viper.scores, diggit.scores, pathway.scores)
}
\arguments{
\item{viper.scores}{numeric Vector}

\item{diggit.scores}{List indexed by type char, with numeric score vectors in [0,R+] for each}

\item{pathway.scores}{List , double indexed by each pathway dataset, 
then with type char. Each points to a numeric score vectors in [0,R+] for each}
}
\value{
a named vector of empirical p-values for each protein/candidate Master Regulator
}
\description{
Implements the conditional Bayes model to combine VIPER scores with diggit 
and pathway scores
}
\keyword{internal}
