% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{convertToOncoprint}
\alias{convertToOncoprint}
\title{Helper function to prep plots for oncoprint}
\usage{
convertToOncoprint(momaObj, cnv.threshold, fCNVs)
}
\arguments{
\item{momaObj}{main results object}

\item{cnv.threshold}{cnv threshold to use for binarizing cnv matrix, option of one number, two numbers or "gistic"}

\item{fCNVs}{vector of functional CNVs if provided}
}
\value{
list of matrices indexed by event type
}
\description{
Helper function to prep plots for oncoprint
}
\keyword{internal}
