% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{empirical.p}
\alias{empirical.p}
\title{Get the empirical p-value from a distribution (vector)}
\usage{
empirical.p(gene.name, x)
}
\arguments{
\item{gene.name}{Character}

\item{x}{named Vector of scores for the distribution}
}
\value{
a numeric p-value between 0 and 1
}
\description{
Get the empirical p-value from a distribution (vector)
}
