% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.gbm.data.r
\docType{data}
\name{example.gbm.mae}
\alias{example.gbm.mae}
\title{Glioblastoma (GBM) Example Dataset}
\format{
An MultiAssayExperiment object with 4 different sets of GBM assays
\describe{
\item{viper}{matrix of viper scores with samples in columns and regulators
 across the rows}
\item{mut}{matrix of samples and genes with potential mutations.
0 for no mutation, 1 for presence of some non-silent mutation}
\item{cnv}{matrix of samples and genes with copy number variant scores}
\item{fusion}{matrix of samples and fusion genes.
0 for no fusion, 1 for presence of fusion}
}
}
\usage{
example.gbm.mae
}
\description{
MultiAssayExperiment Object containing all the genomic assays needed to run
the example code for MOMA
}
\keyword{datasets}
