% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{fcnvAnalysis}
\alias{fcnvAnalysis}
\title{Perform fCNV analysis as provided in the diggit bioconductor package}
\usage{
fcnvAnalysis(momaObj, fCNV.sig = 0.05, fCNV.method = "mi", expression = NULL)
}
\arguments{
\item{momaObj}{main object}

\item{fCNV.sig}{significance threshold to use for determining functionality}

\item{fCNV.method}{method to use, either spearman or mutual information ('mi')}

\item{expression}{expression matrix supplied for the analysis if not directly in momaObj}
}
\value{
Vector of CNV names that are determined to be functionally significant
}
\description{
Perform fCNV analysis as provided in the diggit bioconductor package
}
\keyword{internal}
