% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{fishers.integration}
\alias{fishers.integration}
\title{Fisher's integration}
\usage{
fishers.integration(p)
}
\arguments{
\item{p}{vector of p values}
}
\value{
integrated p value
}
\description{
Helper function to do fisher's integration of set of p values
}
\keyword{internal}
