% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{fit.threshold}
\alias{fit.threshold}
\title{Helper function to do inflection point fitting. 
 Use the heuristic, or if supplied, the definitions file to set the 
 MR threshold
 @param sweep : numeric vector of genomic coverage values, named by -k- threshold
 @param frac : Fraction of coverage to use as a threshold (default .85 = 85 percent)
 @return The -k- integer where coverage is acheived}
\usage{
fit.threshold(sweep, frac = NULL)
}
\description{
Helper function to do inflection point fitting. 
 Use the heuristic, or if supplied, the definitions file to set the 
 MR threshold
 @param sweep : numeric vector of genomic coverage values, named by -k- threshold
 @param frac : Fraction of coverage to use as a threshold (default .85 = 85 percent)
 @return The -k- integer where coverage is acheived
}
