% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.gbm.data.r
\docType{data}
\name{gbm.pathways}
\alias{gbm.pathways}
\title{Glioblastoma (GBM) Pathways}
\format{
A list of lists named "cindy" and "preppi" respectively
\describe{
\item{cindy}{list of regulators, each with a set of modulators and p values
representing their CINDY inferred association}
\item{preppi}{list of regulators, each with a set of potential binding
partners and PREPPi inferred p values for probability of binding}
}
}
\usage{
gbm.pathways
}
\description{
Object containing information about the biological pathways that will be used
in the analysis
}
\keyword{datasets}
