% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gene.map}
\alias{gene.map}
\title{Gene Location Mapping}
\format{
A Data frame with 4 columns
\describe{
\item{Gene.Symbol}{Approved Symbol gene name}
\item{Entrez.IDs}{NCBI Gene ID}
\item{Cytoband}{Chromosome location}
\item{Ensembl}{Ensembl gene ID}
}
 @source \url{https://www.genenames.org/download/custom/}
}
\usage{
gene.map
}
\description{
Table used for converting between different forms of gene information.
Downloaded from HGNC's custom download portal using the "Approved Symbol", 
"NCBI Gene ID", "Chromosome" and "Ensembl Gene ID" curated data options 
and only those with "Approved" status.
Updated December 2019.
}
\keyword{datasets}
