% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{genomicPlotSmall}
\alias{genomicPlotSmall}
\title{Make small genomic plot}
\usage{
genomicPlotSmall(input.df, best.k, tissue.cluster = NULL, null.matrix)
}
\arguments{
\item{input.df}{tissue.coverage.df with mean, k, fraction and unique events.}

\item{best.k}{number of regulators in the checkpoint}

\item{tissue.cluster}{which cluster to look at}

\item{null.matrix}{fractions generated for null model}
}
\value{
output .png
}
\description{
Make small genomic plot
}
\keyword{internal}
