% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{genomicSaturationSummary}
\alias{genomicSaturationSummary}
\title{Merge Genomic Saturation 
 
Updated function to generate summary table of upstream genomics calculations
for plotting and determination of checkpoint}
\usage{
genomicSaturationSummary(coverage.range, topN)
}
\arguments{
\item{coverage.range}{list object generated by sampleEventOverlap function with per sample coverage information}

\item{topN}{number of regulators to consider}
}
\value{
dataframe with summary information of cluster wide event coverage
}
\description{
Merge Genomic Saturation 
 
Updated function to generate summary table of upstream genomics calculations
for plotting and determination of checkpoint
}
\keyword{internal}
