% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{get.best.clustering.supervised}
\alias{get.best.clustering.supervised}
\title{Get the set of optimal clusterings from the Silhouette results. Find the one that optimizes survival}
\usage{
get.best.clustering.supervised(cluster.sweep = NULL, clinical.tibble,
  tissue, progression.free.surv = FALSE)
}
\description{
Get the set of optimal clusterings from the Silhouette results. Find the one that optimizes survival
}
