% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{get.clin}
\alias{get.clin}
\title{Parse the full clinical merged file from GDAC-Firehose and make a dlpyr tibble out of it}
\usage{
get.clin(clin.file = NULL)
}
\description{
Parse the full clinical merged file from GDAC-Firehose and make a dlpyr tibble out of it
}
