% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{get.coverage}
\alias{get.coverage}
\title{Get coverage of interactions}
\usage{
get.coverage(momaObj, cMR.ranking, viper.samples, topN = 100,
  mutation.filter = NULL)
}
\arguments{
\item{momaObj}{: numeric vector with cluster membership, names are samples}

\item{cMR.ranking}{: a vector entrez IDs, in order}

\item{viper.samples}{: calculate the genomic coverage only for these sample}
}
\description{
Get coverage of interactions
}
