% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{get.diggit.empiricalQvalues}
\alias{get.diggit.empiricalQvalues}
\title{Compute the empirical q-values of each genomic-event/VIPER gene pair against the background
distribution of associations with a given set of 'null' VIPER genes (i.e. low activity TFs)}
\usage{
get.diggit.empiricalQvalues(vipermat, nes, null.TFs, alt = "greater")
}
\arguments{
\item{vipermat}{viper inferences matrix, samples are columns, rows are TF entrez gene IDs}

\item{nes}{scores for each mutation (rows) against each TF (columns)}

\item{null.TFs}{low-importance TFs used to calculate null distributions}
}
\description{
Compute the empirical q-values of each genomic-event/VIPER gene pair against the background
distribution of associations with a given set of 'null' VIPER genes (i.e. low activity TFs)
}
