% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{get.empirical.qvals}
\alias{get.empirical.qvals}
\title{get.empirical.qvals}
\usage{
get.empirical.qvals(test.statistics, null.statistics,
  alternative = "both")
}
\arguments{
\item{test.statistics}{P-values generated from the test comparisons}

\item{null.statistics}{P-values generated under the null (permutation) model}

\item{alternative}{Optional : 1 or 2 tails used to generate the p-value (default='both')}
}
\description{
get.empirical.qvals
}
