% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{get.pvals.matrix}
\alias{get.pvals.matrix}
\title{Utility function}
\usage{
get.pvals.matrix(corrected.scores)
}
\arguments{
\item{corrected.scores}{- corrected p-values processed by 'qvals' package}
}
\description{
Utility function
}
