% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{get_clin}
\alias{get_clin}
\title{get_clin Parse the full clinical merged file from GDAC-Firehose and make a dlpyr tibble out of it}
\usage{
get_clin(clin.file = NULL)
}
\arguments{
\item{clin.file}{Clinical Merge file from TCGA}
}
\value{
Returns a tibble data.frame of the clinical data. Rows are patient.bcr_patient_barcode IDs
}
\description{
get_clin Parse the full clinical merged file from GDAC-Firehose and make a dlpyr tibble out of it
}
