% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{getBestClusteringSupervised}
\alias{getBestClusteringSupervised}
\title{Get best clustering supervised}
\usage{
getBestClusteringSupervised(
  search.results,
  top.sol,
  survival.data,
  progression.free
)
}
\arguments{
\item{search.results}{results object from clusterRange with all the information
for each possible clustering solution}

\item{top.sol}{the top solution from the first pass of the analysis}

\item{survival.data}{dataframe with survival data for each sample}

\item{progression.free}{T/F about whether to use progression free instead of overall survival}
}
\value{
best clustering solution k
}
\description{
This function calculates the best clustering based on optimizing for survival
separation in the event there are statistically equivalent solutions
}
\keyword{internal}
