% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{getCoverage}
\alias{getCoverage}
\title{Get coverage of interactions}
\usage{
getCoverage(
  momaObj,
  cMR.ranking,
  viper.samples,
  topN = 100,
  mutation.filter = NULL,
  verbose = FALSE
)
}
\arguments{
\item{momaObj}{A numeric vector with cluster membership, names are samples}

\item{cMR.ranking}{A vector entrez IDs, in order}

\item{viper.samples}{Calculate the genomic coverage only for these sample}

\item{topN}{Compute coverage for only the top -N- Master Regulators}

\item{mutation.filter}{Retain only mutation events in this (positive) list}
}
\value{
A list of lists, indexed by sample name, with coverage statistics
for each sample
}
\description{
Get coverage of interactions
}
\keyword{internal}
