% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{getDataFrame}
\alias{getDataFrame}
\title{Helper function to get data frame for bar plot plot.events function}
\usage{
getDataFrame(
  data,
  highlight.genes,
  genomeBand_2_gene,
  max.muts = 10,
  max.cnv = 5
)
}
\arguments{
\item{data}{data.frame with $type, $id, $Freq per event}

\item{highlight.genes}{genes to look for in mutations/cnv lists (if looking
for specific genes because of prior knowledge)}

\item{genomeBand_2_gene}{mapping of genomic location IDs to gene name: 
vector of HUGO gene ids, named by genomic loci}

\item{max.muts}{maximum number of mutations to get per sample, default is 10}

\item{max.cnv}{maximum number of cnvs to per sample, default is 5}
}
\value{
ordered data frame with each genomic event and it's frequency
}
\description{
Helper function to get data frame for bar plot plot.events function
}
\keyword{internal}
