% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{getInflection}
\alias{getInflection}
\title{Get Saturation Inflection Point}
\usage{
getInflection(fractions, clus.id)
}
\arguments{
\item{fractions}{vector of values representing fraction of events covered by that number of top regulators}

\item{clus.id}{cluster being evaluated}
}
\value{
number representative of regulators in the checkpoint
}
\description{
Function to calculate checkpoint based on derivative of coverage fraction values
}
\keyword{internal}
