% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{getMaxEvents}
\alias{getMaxEvents}
\title{Function to subset final matrix to max desired for plotting}
\usage{
getMaxEvents(mat.to.plot, max.events)
}
\arguments{
\item{mat.to.plot}{matrix of events to be plotted}

\item{max.events}{maximum number of events to plot for the oncoplots}
}
\value{
matrix with reduced overall events
}
\description{
Function to subset final matrix to max desired for plotting
}
\keyword{internal}
