% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.interactions.r
\name{getPvalsMatrix}
\alias{getPvalsMatrix}
\title{Utility function}
\usage{
getPvalsMatrix(corrected.scores)
}
\arguments{
\item{corrected.scores}{- corrected p-values processed by 'qvals' package}
}
\value{
A matrix of p-values for scores between genes/events (rows) and TFs (columns)
}
\description{
Utility function
}
