% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma_arguments.R
\name{group_lasso}
\alias{group_lasso}
\alias{moma_grplasso}
\title{Group LASSO}
\usage{
moma_grplasso(..., g, non_negative = FALSE, ..., lambda = 0,
  select_scheme = "g")
}
\arguments{
\item{...}{Forces users to specify all arguments by name.}

\item{g}{A vector of integer or characters, or a factor itself. It gets transformed
to factor eventually to indicate grouping.}

\item{non_negative}{A Boolean value. Set it to \code{TRUE} to add non-negativity
constraint.}

\item{lambda}{A vector containing penalty values}

\item{select_scheme}{A char being either "b" (nested BIC search) or "g" (grid search).

MoMA provides a flexible framework for regularized multivariate analysis
with several tuning parameters for different forms of regularization.
To assist the user in selecting these parameters (\code{alpha_u},
\code{alpha_v}, \code{lambda_u}, \code{lambda_v}), we provide
two selection modes: grid search ("g") and nested BIC search ("b").
Grid search means we solve the problem
for all combinations of parameter values provided by the user.

To explain nested BIC search, we need to look into how the algorithm runs.
To find an (approximate) solution to a penalized SVD (Singular Value Decomposition) problem is to solve two
penalized regression problems iteratively. Let's call them problem u and problem v, which give
improving estimates of the right singular vector, \emph{u}, and the left singular vector, \emph{v}, respectively.
For each regression problem, we can select the optimal parameters
based on BIC.

The nested BIC search is essentially two 2-D searches. We start from SVD solutions, and then find the optimal
parameters for problem u, given current estimate of \emph{v}. Using the result from previous step, update
current estimate of \emph{u}, and then do the same thing for problem v,
that is, to find the optimal parameters for problem v given current estimate of \emph{u}. Repeat
the above until convergence or the maximal number of iterations has been reached.

Users are welcome to refer to section 3.1: Selection of Regularization Parameters
in the paper cited below.}
}
\value{
A \code{moma_sparsity_type} object, which is a list containing the values of \code{non_negative}
and \code{g}.
}
\description{
Use this function to set the penalty function to group lasso
\deqn{\lambda \sum_{g \in Group} \| x_g \|_1,}
where \eqn{\lambda} is set by the \code{lambda} argument below, \eqn{x_g} is
the vector comprised of elements of \eqn{x} picked out by the indices set \eqn{g}.
}
\references{
Yuan, Ming, and Yi Lin. "Model Selection and Estimation in Regression
with Grouped Variables." Journal of the Royal Statistical Society:
Series B (Statistical Methodology) 68.1 (2006): 49-67. \doi{10.1111/j.1467-9868.2005.00532.x}.
}
