% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{make.coverage.df}
\alias{make.coverage.df}
\title{Helper function for making the coverage dataframe (might)}
\usage{
make.coverage.df(coverage.list, cutoff)
}
\arguments{
\item{coverage.list}{: List indexed by sample name, contains $mut, $fus, $amp, $del interactions}

\item{cutoff}{: number of regulators to include}
}
\value{
dataframe with each sample and which events are captured by the checkpoint mrs
}
\description{
Helper function for making the coverage dataframe (might)
}
