% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{makeCoverageDf}
\alias{makeCoverageDf}
\title{Helper function for making the coverage dataframe (might)}
\usage{
makeCoverageDf(coverage.list, cutoff)
}
\arguments{
\item{coverage.list}{: List indexed by sample name, contains $mut, $fus, $amp, $del interactions}

\item{cutoff}{: number of regulators to include}
}
\value{
dataframe with each sample and which events are captured by the checkpoint mrs
}
\description{
Helper function for making the coverage dataframe (might)
}
