% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{makeCoverageDfNew}
\alias{makeCoverageDfNew}
\title{Helper function for making the coverage dataframe for version 2 results}
\usage{
makeCoverageDfNew(coverage.list, cutoff)
}
\arguments{
\item{coverage.list}{: List indexed by sample name, 
contains mut/fus/amp/del interactions}

\item{cutoff}{: number of regulators to include}
}
\value{
summary data frame with which events are 
captured by the checkpoint mrs
}
\description{
Helper function for making the coverage dataframe for version 2 results
}
\keyword{internal}
