% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.functions.R
\name{map_hugo}
\alias{map_hugo}
\title{Convert from hugo gene names to entrez ids}
\usage{
map_hugo(hugo.ids)
}
\arguments{
\item{hugo.ids}{: vector of hugo gene names, requires hugo2entrez to be loaded}
}
\value{
: vector of entrez ids
}
\description{
Convert from hugo gene names to entrez ids
}
