% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.functions.R
\name{mapEntrez}
\alias{mapEntrez}
\title{Convert from entrez ids to hugo gene names}
\usage{
mapEntrez(entrez.ids)
}
\arguments{
\item{entrez.ids}{: vector of entrez ids requires hugo2entrez to be loaded}
}
\value{
: vector of hugo gene names
}
\description{
Convert from entrez ids to hugo gene names
}
\examples{
map_entrez(c("29974", "5728"))
}
\seealso{
\code{\link[moma]{map_hugo}}
}
