% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.functions.R
\name{mapHugo}
\alias{mapHugo}
\title{Convert from hugo gene names to entrez ids}
\usage{
mapHugo(hugo.ids)
}
\arguments{
\item{hugo.ids}{: vector of hugo gene names, requires hugo2entrez to be loaded}
}
\value{
: vector of entrez ids
}
\description{
Convert from hugo gene names to entrez ids
}
\examples{
mapHugo(c("A1CF","PTEN"))
}
\seealso{
\code{\link[MOMA]{mapEntrez}}
}
