% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.functions.r
\name{mapScores.cnvBand}
\alias{mapScores.cnvBand}
\title{mapping: a named vector of genomic locations/cytoband IDs. 
names are the gene names for each--i.e. a many to one mapping from HUGO or entrez IDs to
cytoband location}
\usage{
mapScores.cnvBand(mapping, diggit.interactions, from.p = FALSE,
  pos.nes.only = TRUE)
}
\description{
mapping: a named vector of genomic locations/cytoband IDs. 
names are the gene names for each--i.e. a many to one mapping from HUGO or entrez IDs to
cytoband location
}
