% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.old.r
\name{mapScoresCnvBand}
\alias{mapScoresCnvBand}
\title{Map scores to cytoband location}
\usage{
mapScoresCnvBand(
  mapping,
  diggit.interactions,
  from.p = FALSE,
  pos.nes.only = TRUE
)
}
\arguments{
\item{mapping}{a named vector of genomic locations/cytoband IDs. 
names are the gene names for each--i.e. a many to one mapping from HUGO or entrez IDs to
cytoband location}

\item{diggit.interactions}{list indexed by MR/TF name in Entrez Space}

\item{from.p}{DIGGIT interactions are in p-value format instead of z-score (default=FALSE)}

\item{pos.nes.only}{Only consider positive associations with NES scores (default=TRUE)
each points to a named vector of NES / z-scores associated with entrez IDs for each interacting event.}
}
\value{
A list of input scores, now named by cytoband location
}
\description{
Map scores to cytoband location
}
