% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{merge.genomicSaturation}
\alias{merge.genomicSaturation}
\title{Create data frame from coverage data, including number of total events 'covered' and unique events}
\usage{
\method{merge}{genomicSaturation}(coverage.range, topN)
}
\description{
Create data frame from coverage data, including number of total events 'covered' and unique events
}
