% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{mergeDataBySubtype}
\alias{mergeDataBySubtype}
\title{Create data frame from coverage data, including number of total events 
'covered' and unique events}
\usage{
mergeDataBySubtype(genomic.saturation, sample.clustering, topN = 100, new)
}
\arguments{
\item{genomic.saturation}{: data from genomic saturation function}

\item{sample.clustering}{: clustering vector with sample names and 
cluster designations}

\item{topN}{: number of regulators to look through. default is 100}

\item{new}{T/F based on whether or not using results from version 1 or 2 of the pipeline}
}
\value{
dataframe with coverage data for genomic events
}
\description{
Create data frame from coverage data, including number of total events 
'covered' and unique events
}
\keyword{internal}
