% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{mergeDataNew}
\alias{mergeDataNew}
\title{Helper function for mergeDataBySubtype}
\usage{
mergeDataNew(coverage.range, topN)
}
\arguments{
\item{coverage.range}{: genomic saturation for a particular subtype}

\item{topN}{: max number of top regulators to search through}
}
\value{
dataframe with coverage data for genomic events
}
\description{
Helper function for mergeDataBySubtype
}
\keyword{internal}
