% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{mergeDuplicates}
\alias{mergeDuplicates}
\title{Function to merge duplicates before oncoprint plotting}
\usage{
mergeDuplicates(mat.to.plot)
}
\arguments{
\item{mat.to.plot}{matrix of events to be plotted}
}
\value{
matrix with merged duplicate rows
}
\description{
Function to merge duplicates before oncoprint plotting
}
\keyword{internal}
