% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{mergeLists}
\alias{mergeLists}
\title{Helper function}
\usage{
mergeLists(l1, l2)
}
\arguments{
\item{l1}{list 1}

\item{l2}{list 2}
}
\value{
single merged list
}
\description{
Helper function
}
\keyword{internal}
