% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.match.r
\docType{class}
\name{model_viperMatcher-class}
\alias{model_viperMatcher-class}
\alias{model_viperMatcher}
\title{Encapsulates the logic needed to match patient samples to a set of cell-line or other models.}
\arguments{
\item{patient.viper}{- VIPER matrix of patient data (columns = patients; rows = entrez IDs)}

\item{model.viper}{- VIPER matrix of model data (columns = model names; rows = entrez IDs)}

\item{checkpoints}{- list of checkpoint definitions}

\item{clusters}{- list of cluster definitions}

\item{checkpoints}{: a list with each index containing a checkpoint definition (entrezIDS) to match 
against a set of model system VIPER scores}

\item{model.vipermat}{: viper matrix for the model system being tested. Columns are samples, rows proteins

Plot the viper matrix of the best matching cell lines against the 

Plot the viper matrix of the best matching cell lines against the MEAN of each cluster}
}
\description{
Encapsulates the logic needed to match patient samples to a set of cell-line or other models.
}

