% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{moma.constructor}
\alias{moma.constructor}
\title{MOMA Constructor}
\usage{
moma.constructor(viper, mut, cnv, fusions, pathways,
  gene.blacklist = NULL, output.folder = NULL,
  gene.loc.mapping = NULL)
}
\arguments{
\item{mut}{: an indicator matrix (0/1) of mutation events with samples as columns and genes as rows}

\item{cnv}{: a matrix of CNV scores (typically SNP6 array scores from TCGA) with samples as columns and genes as rows}

\item{fusions}{: an indicator matrix (0/1) of fusion events with samples as columns and genes as rows}

\item{pathways}{: a named list of lists. Each named list represents interactions between proteins (keys) and their associated partners}

\item{cytoband.mapping}{: vector of band locations, names are Entrez IDs}
}
\value{
an instance of class momaRunner
}
\description{
MOMA Constructor
}
