\name{moma}
\alias{moma}
\title{Launch the moma Shiny application}
\usage{
moma(launch.browser = interactive(), ...)
}
\arguments{
  \item{launch.browser}{Whether to launch the app in a browser.}

  \item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Invisibly returns the running Shiny app.
}
\description{
Launch the moma Shiny application.
}
