% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{moma_constructor}
\alias{moma_constructor}
\title{MOMA Constructor
see vignette for more information on how to set up and run the MOMA object}
\usage{
moma_constructor(
  viper,
  mut,
  cnv,
  fusions,
  pathways,
  gene.blacklist = NA_character_,
  output.folder = NA_character_,
  gene.loc.mapping = gene.map
)
}
\arguments{
\item{viper}{VIPER protein activity matrix with samples as columns and rows as protein IDs}

\item{mut}{An indicator matrix (0/1) of mutation events with samples as columns and genes as rows}

\item{cnv}{A matrix of CNV scores (typically SNP6 array scores from TCGA) with samples as columns and genes as rows}

\item{fusions}{An indicator matrix (0/1) of fusion events with samples as columns and genes as rows}

\item{pathways}{A named list of lists. Each named list represents interactions between proteins (keys) and their associated partners}

\item{gene.blacklist}{A vector of genes to exclude from mutational/CNV/fusion analysis}

\item{output.folder}{Location to store output and intermediate results}

\item{gene.loc.mapping}{A data.frame of band locations and Entrez IDs}
}
\value{
an instance of class momaRunner
}
\description{
MOMA Constructor
see vignette for more information on how to set up and run the MOMA object
}
\examples{
library(moma.gbmexample) 
data("gbm.example")
momaObj <- moma_constructor(gbm.example$vipermat, 
gbm.example$rawsnp, 
gbm.example$rawcnv, 
gbm.example$fusions, 
pathways = list(cindy=gbm.example$cindy, preppi=gbm.example$preppi), 
gene.blacklist=gbm.example$mutSig)

}
