% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{moma_session_info}
\alias{moma_session_info}
\title{Session Info used for Bug Reporting}
\usage{
moma_session_info()
}
\value{
None. Called for side-effects (printed to the screen)
       only.
}
\description{
A helper function which prints information useful in
bug reports.
}
