% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma_sfcca.R
\name{moma_sfcca}
\alias{moma_sfcca}
\alias{moma_scca}
\alias{moma_twscca}
\alias{moma_fcca}
\alias{moma_twfcca}
\title{Sparse and functional CCA}
\usage{
moma_sfcca(X, ..., Y, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_scca(X, ..., Y, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_twscca(X, ..., Y, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_fcca(X, ..., Y, center = TRUE, scale = FALSE,
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_twfcca(X, ..., Y, center = TRUE, scale = FALSE,
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)
}
\arguments{
\item{X, Y}{A data matrix, each row representing a sample, and each column a feature.}

\item{...}{Force users to specify arguments by names.}

\item{center}{A logical value indicating whether the variables should be shifted to be zero centered.
Defaults to \code{TRUE}.}

\item{scale}{A logical value indicating whether the variables should be scaled to have unit variance.
Defaults to \code{FALSE}.}

\item{x_sparse, y_sparse}{An object of class inheriting from "\code{moma_sparsity_type}". Most conveniently
specified by functions described in \code{\link{moma_sparsity_options}}. It specifies the type of sparsity-inducing
penalty function used in the model. Note that for \code{moma_scca}, these two parameters must not be
specified at the same time. For \code{moma_fcca} and \code{moma_twfcca}, they must not be specified.}

\item{x_smooth, y_smooth}{An object of class inheriting from "\code{moma_smoothness_type}". Most conveniently
specified by functions described in \code{moma_smoothness}. It specifies the type of smoothness
 terms used in the model. Note that for \code{moma_fcca}, these two parameters must not be
specified at the same time. For \code{moma_scca} and \code{moma_twscca}, they must not be specified.}

\item{pg_settings}{An object of class inheriting from "\code{moma_pg_settings}". Most conviently
specified by functions described in \code{\link{moma_pg_settings}}. It specifies the type of algorithm
used to solve the problem, acceptable level of precision, and the maximum number of iterations allowed.}

\item{max_bic_iter}{A positive integer. Defaults to 5. The maximum number of iterations allowed
in nested greedy BIC selection scheme.}

\item{rank}{A positive integer. Defaults to 1. The maximal rank, i.e., maximal number of principal components to be used.}
}
\description{
\code{moma_sfcca} creates an \code{SFCCA} R6 object and returns it. Type \code{?CCA_deflation} for
description of problem formulation and deflation scheme.

\code{moma_scca} is a function for performing one-way sparse CCA.

\code{moma_twscca} is a function for performing two-way sparse CCA.

\code{moma_fcca} is a function for performing one-way functional CCA.

\code{moma_twfcca} is a function for performing two-way functional CCA.
}
\section{Functions}{
\itemize{
\item \code{moma_scca}: a function for performing one-way sparse CCA.

\item \code{moma_twscca}: a function for performing two-way sparse CCA

\item \code{moma_fcca}: a function for performing one-way functional CCA

\item \code{moma_twfcca}: a function for performing two-way functional CCA
}}

