% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma_sflda.R
\name{moma_sflda}
\alias{moma_sflda}
\alias{moma_slda}
\alias{moma_twslda}
\alias{moma_flda}
\alias{moma_twflda}
\title{Sparse and functional LDA}
\usage{
moma_sflda(X, ..., Y_factor, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_slda(X, ..., Y_factor, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_twslda(X, ..., Y_factor, center = TRUE, scale = FALSE,
  x_sparse = moma_empty(), y_sparse = moma_empty(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_flda(X, ..., Y_factor, center = TRUE, scale = FALSE,
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)

moma_twflda(X, ..., Y_factor, center = TRUE, scale = FALSE,
  x_smooth = moma_smoothness(), y_smooth = moma_smoothness(),
  pg_settings = moma_pg_settings(), max_bic_iter = 5, rank = 1)
}
\arguments{
\item{X}{A data matrix, each row representing a sample, and each column a feature.}

\item{...}{Force users to specify arguments by names.}

\item{Y_factor}{A factor representing which group a sample belongs to.}

\item{center}{A logical value indicating whether the variables should be shifted to be zero centered.
Defaults to \code{TRUE}.}

\item{scale}{A logical value indicating whether the variables should be scaled to have unit variance.
Defaults to \code{FALSE}.}

\item{x_sparse}{An object of class inheriting from "\code{moma_sparsity_type}". Most conveniently
specified by functions described in \code{\link{moma_sparsity_options}}. It specifies the type of sparsity-inducing
penalty function used in the model. Note that for \code{moma_scca}, these two parameters must not be
specified at the same time. For \code{moma_fcca} and \code{moma_twfcca}, they must not be specified.}

\item{y_sparse}{An object of class inheriting from "\code{moma_sparsity_type}". Most conveniently
specified by functions described in \code{\link{moma_sparsity_options}}. It specifies the type of sparsity-inducing
penalty function used in the model. Note that for \code{moma_scca}, these two parameters must not be
specified at the same time. For \code{moma_fcca} and \code{moma_twfcca}, they must not be specified.}

\item{x_smooth}{An object of class inheriting from "\code{moma_smoothness_type}". Most conveniently
specified by functions described in \code{moma_smoothness}. It specifies the type of smoothness
 terms used in the model. Note that for \code{moma_fcca}, these two parameters must not be
specified at the same time. For \code{moma_scca} and \code{moma_twscca}, they must not be specified.}

\item{y_smooth}{An object of class inheriting from "\code{moma_smoothness_type}". Most conveniently
specified by functions described in \code{moma_smoothness}. It specifies the type of smoothness
 terms used in the model. Note that for \code{moma_fcca}, these two parameters must not be
specified at the same time. For \code{moma_scca} and \code{moma_twscca}, they must not be specified.}

\item{pg_settings}{An object of class inheriting from "\code{moma_pg_settings}". Most conviently
specified by functions described in \code{\link{moma_pg_settings}}. It specifies the type of algorithm
used to solve the problem, acceptable level of precision, and the maximum number of iterations allowed.}

\item{max_bic_iter}{A positive integer. Defaults to 5. The maximum number of iterations allowed
in nested greedy BIC selection scheme.}

\item{rank}{A positive integer. Defaults to 1. The maximal rank, i.e., maximal number of principal components to be used.}
}
\description{
\code{moma_sflda} creates an \code{SFLDA} R6 object and returns it.

\code{moma_slda} is a function for performing one-way sparse LDA.

\code{moma_twslda} is a function for performing two-way sparse LDA.

\code{moma_flda} is a function for performing one-way functional LDA.

\code{moma_twflda} is a function for performing two-way functional LDA.
}
\section{Functions}{
\itemize{
\item \code{moma_slda}: a function for performing one-way sparse LDA

\item \code{moma_twslda}: a function for performing two-way sparse LDA

\item \code{moma_flda}: a function for performing one-way functional LDA

\item \code{moma_twflda}: a function for performing two-way functional LDA
}}

