% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma_arguments.R
\name{moma_smoothness}
\alias{moma_smoothness}
\title{Smoothness-inducing Term}
\usage{
moma_smoothness(Omega = NULL, ..., alpha = 0, select_scheme = "g")
}
\arguments{
\item{Omega}{A matrix of appropriate size. A common choice is the second difference matrix.
See \code{\link{second_diff_mat}}.}

\item{...}{Force users to specify arguments by names.}

\item{alpha}{A vector containing penalty values}

\item{select_scheme}{A char being either "b" (nested BIC search) or "g" (grid search).

MoMA provides a flexible framework for regularized multivariate analysis
with several tuning parameters for different forms of regularization.
To assist the user in selecting these parameters (\code{alpha_u},
\code{alpha_v}, \code{lambda_u}, \code{lambda_v}), we provide
two selection modes: grid search ("g") and nested BIC search ("b").
Grid search means we solve the problem
for all combinations of parameter values provided by the user.

To explain nested BIC search, we need to look into how the algorithm runs.
To find an (approximate) solution to a penalized SVD (Singular Value Decomposition) problem is to solve two
penalized regression problems iteratively. Let's call them problem u and problem v, which give
improving estimates of the right singular vector, \emph{u}, and the left singular vector, \emph{v}, respectively.
For each regression problem, we can select the optimal parameters
based on BIC.

The nested BIC search is essentially two 2-D searches. We start from SVD solutions, and then find the optimal
parameters for problem u, given current estimate of \emph{v}. Using the result from previous step, update
current estimate of \emph{u}, and then do the same thing for problem v,
that is, to find the optimal parameters for problem v given current estimate of \emph{u}. Repeat
the above until convergence or the maximal number of iterations has been reached.

Users are welcome to refer to section 3.1: Selection of Regularization Parameters
in the paper cited below.}
}
\description{
This function specifies the value of the \code{u_smooth,v_smooth} arguments in the
 \code{moma_*pca} series of functions, and the \code{x_smooth,y_smooth} arguments
in the \code{moma_*cca} and \code{moma_*lda} series of functions.
}
