% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{nullQuantiles}
\alias{nullQuantiles}
\title{Function to get lower and upper quantiles for the null model}
\usage{
nullQuantiles(null.matrix)
}
\arguments{
\item{null.matrix}{table of fractions}
}
\value{
list with upper and lower quantile
}
\description{
Function to get lower and upper quantiles for the null model
}
\keyword{internal}
