% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{oncoprintParams}
\alias{oncoprintParams}
\title{Function to save oncoprint parameters to a list
Also possible to supply new oncoprint parameters via this function}
\usage{
oncoprintParams(oncoprint.params, final.mat, cnv.threshold)
}
\arguments{
\item{oncoprint.params}{optional list of oncoprint parameters to supply if 
additional types of events are added/ if different parameters want to be supplied}

\item{final.mat}{matrix to be plotted}

\item{cnv.threshold}{: value cut off for binarizing cnvs, option of one number, two numbers or "gistic"}
}
\value{
list of functions/parameters
}
\description{
Function to save oncoprint parameters to a list
Also possible to supply new oncoprint parameters via this function
}
\keyword{internal}
