% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{oncoprintPlotNew}
\alias{oncoprintPlotNew}
\title{Function to plot genomic events in the style of oncoPrint/cBioPortal using new data structures}
\usage{
oncoprintPlotNew(
  summary.table,
  clus.event.mats,
  clus.genomic.saturation,
  cutoff,
  important.genes,
  max.events,
  k,
  cytoband.collapse,
  oncoprint.params,
  cnv.threshold
)
}
\arguments{
\item{summary.table}{summary table of all events in a particular subtype}

\item{clus.event.mats}{list of submatrices of events for this cluster, indexed by type}

\item{clus.genomic.saturation}{results of genomic saturation analysis for this cluster}

\item{cutoff}{number of regulators to include}

\item{important.genes}{well known genes to highlight in the analysis}

\item{max.events}{maximum number of events to plot for the oncoplots}

\item{k}{current cluster number}

\item{cytoband.collapse}{T/F regarding whether or not the genomic saturation
analysis was done with collapsing the CNVs to cytobands}

\item{oncoprint.params}{optional list of oncoprint parameters to supply if 
additional types of events are added / if different parameters want to be supplied}

\item{cnv.threshold}{: value cut off for binarizing cnvs, option of one number, two numbers or "gistic"}
}
\value{
oncoprint event plot
}
\description{
Function to plot genomic events in the style of oncoPrint/cBioPortal using new data structures
}
\keyword{internal}
