% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{overrep.analysis}
\alias{overrep.analysis}
\title{Over-representation Test for genomic events}
\usage{
overrep.analysis(mat, inCluster.samples, min.events.per.cluster)
}
\arguments{
\item{mat}{binary events matrix}

\item{inCluster.samples}{names of samples in the cluster to be tested}

\item{min.events.per.cluster}{minimum number of times that event must occur in that cluster to be countable}
}
\value{
names of genes that had marginal over-representation
}
\description{
Do a proportion test for representation of a particular genomic event in a cluster
}
\keyword{internal}
