% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma_sfpca.R
\name{PCA_deflation}
\alias{PCA_deflation}
\title{Deflation Schemes for PCA}
\description{
In \code{MoMA} three deflation schemes are provided for PCA.
Using terminology in the reference, they are Hotelling's deflation,
two-way projection deflation, and Schur complement deflation.
}
\details{
See the parameter \code{deflation_scheme} argument in the function
\code{moma_sfpca}. Also refer to the reference below
for theoretical properties.
}
\references{
Michael Weylandt. "Multi-Rank Sparse and Functional PCA: Manifold Optimization and
Iterative Deflation Techniques." arXiv:1907.12012v1, 2019.
}
