% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{plotEvents}
\alias{plotEvents}
\title{Plot barchart of genomic events}
\usage{
plotEvents(
  summary.vec,
  highlight.genes = NULL,
  genomeBand_2_gene = NULL,
  samples.total,
  max.muts = 10,
  max.cnv = 5
)
}
\arguments{
\item{summary.vec}{: named vector of the counts, named 'Event name':'Type'
where type is 'mut', 'amp', 'del', 'fus'. Mutations are in Entrez ID
Amp/Deletion CNV events are in genomic band location}

\item{highlight.genes}{: well known genes to highlight in the analysis in}

\item{genomeBand_2_gene}{: mapping of genomic location IDs to gene name: 
vector of HUGO gene ids, named by genomic loci}

\item{samples.total}{: number of samples in the subtype, used to calculate percentages}

\item{max.muts}{: maximum number of mutations to get per sample, default is 10}

\item{max.cnv}{: maximum number of cnvs to per sample, default is 5}
}
\value{
plot object
}
\description{
Plot barchart of genomic events
}
\keyword{internal}
