% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{rankNormalize}
\alias{rankNormalize}
\title{Rank Normalize P Values}
\usage{
rankNormalize(interactions.df)
}
\arguments{
\item{interacterations.df}{dataframe off all the interactions}
}
\value{
dataframe will adjusted p values
}
\description{
This function uses an empirical cumulative distribution function to rank 
normalize all the p-values from a particular test
}
\keyword{internal}
