% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{reaNULL}
\alias{reaNULL}
\title{This function generates the NULL model function, which computes the normalized enrichment score and associated p-value}
\usage{
reaNULL(regulon, minsize = 1, maxsize = Inf)
}
\arguments{
\item{regulon}{A list with genomic features as its names and samples as its entries}

\item{minsize}{Minimum number of event (or size of regulon) to calculate the model with}
}
\value{
A list of functions to compute NES and p-value
}
\description{
This function generates the NULL model function, which computes the normalized enrichment score and associated p-value
}
