% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{sample.overlap}
\alias{sample.overlap}
\title{The core function to compute which sample-specific alterations overlap with genomic events that are explained 
via DIGGIT.}
\usage{
sample.overlap(momaObj, viper.samples, selected.tfs, interaction.map,
  cnv.threshold = 0.5, mutation.filter = NULL, verbose = FALSE,
  idx.range = NULL)
}
\arguments{
\item{momaObj}{Object reference of momaRunner class}

\item{viper.samples}{Sample vector to restrict sample-specific analysis to}

\item{selected.tfs}{Transcription factors being analyzed}

\item{interaction.map}{List object of events 'covered' by the supplied interactions of type mut/amp/del/fus}

\item{cnv.threshold}{Numeric absolute value to threshold SNP6 and/or GISTIC or other CNV scores at. Above that absolute value is considered a positive event.}

\item{mutation.filter}{A vector of whitelisted mutation events, in entrez gene IDs}

\item{verbose}{Output status during the run (default=FALSE)}

\item{idx.range}{Number of tfs to check for genomic saturation calculation, default is 1253}
}
\description{
The core function to compute which sample-specific alterations overlap with genomic events that are explained 
via DIGGIT.
}
