% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.R
\name{sampleEventOverlap}
\alias{sampleEventOverlap}
\title{Sample Event Overlap}
\usage{
sampleEventOverlap(
  momaObj,
  viper.samples,
  selected.tfs,
  interaction.df,
  cytoband.collapse = T,
  topN = 100
)
}
\arguments{
\item{momaObj}{main object holding all the previously calculated results}

\item{viper.samples}{vector of samples to be analyzed}

\item{selected.tfs}{top ranking tfs that are active in the samples under consideration}

\item{interaction.df}{dataframe of all interactions associated with the regulators}

\item{cytoband.collapse}{T/F of whether to consolidate cnvs in the same cytoband to be considered as one event}

\item{topN}{how many regulators to consider for the saturation analysis}
}
\value{
a list obj with all the coverage information for each sample analyzed
}
\description{
Updated function to compute which sample-specific alterations overlap with 
genomic events that are explained via DIGGIT/upstream of top regulators.
}
\keyword{internal}
