% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{saturationNull}
\alias{saturationNull}
\title{Function to generate null model for the saturation curves}
\usage{
saturationNull(
  momaObj,
  viper.samples,
  tissue.cluster,
  topN,
  cytoband.collapse = T,
  interaction.p = 0.05,
  aqtl.p = 1,
  new = F
)
}
\arguments{
\item{momaObj}{momaObj}

\item{viper.samples}{samples to consider}

\item{tissue.cluster}{which cluster to look at}

\item{topN}{number of MRs to take into account}

\item{cytoband.collapse}{T/F regarding whether or not the genomic saturation analysis was done with collapsing the CNVs to cytobands}

\item{interaction.p}{pvalue threshold to use for MR-event interactions}

\item{aqtl.p}{pvalue threshold to use for aQTL scores}

\item{new}{T/F based on whether or not using results from version 1 or 2 of the pipeline}
}
\value{
null means object
}
\description{
Function to generate null model for the saturation curves
}
\keyword{internal}
