% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{second_diff_mat}
\alias{second_diff_mat}
\title{Second difference matrix}
\usage{
second_diff_mat(n)
}
\arguments{
\item{n}{An integer. The size of the returned matrix.}
}
\description{
This function returns a second difference matrix of size \eqn{n}.
}
