% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma_arguments.R
\name{select_scheme}
\alias{select_scheme}
\title{Introduction to selection schemes in MoMA}
\description{
Please see the description of the argument \code{select_scheme} in
\code{\link{moma_sparsity_options}}. The \code{select_scheme} argument
presents in functions listed in \code{\link{moma_sparsity_options}}, and the function
\code{\link{moma_smoothness}}.
}
