% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{sREA}
\alias{sREA}
\title{Simple one-tail rank based enrichment analysis sREA
(for cluster analysis)}
\usage{
sREA(signatures, groups)
}
\arguments{
\item{signatures}{Numeric matrix of signatures}

\item{groups}{List containing the groups as vectors of sample names}
}
\value{
Matrix of Normalized Enrichment Zcores
}
\description{
This function performs simple 1-tail rank based enrichment analysis
}
\keyword{internal}
