% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{stouffer.integrate}
\alias{stouffer.integrate}
\title{dispatch method for either CNV location corrected or SNV}
\usage{
stouffer.integrate(interactions, cytoband.map = NULL)
}
\description{
dispatch method for either CNV location corrected or SNV
}
