% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{subset.list.interactions}
\alias{subset.list.interactions}
\title{helper function: subset a list to the set of keys supplied
return the names of interactions with positive values, in a list structure}
\usage{
\method{subset}{list.interactions}(int.l, keys)
}
\description{
helper function: subset a list to the set of keys supplied
return the names of interactions with positive values, in a list structure
}
