% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{subsetEventPlots}
\alias{subsetEventPlots}
\title{Function to oncoprint submat to plot
Helper function to deal with getting best event for amps/dels}
\usage{
subsetEventPlots(
  summary.table,
  clus.event.mats,
  etype,
  full.coverage.df,
  cytoband.collapse
)
}
\arguments{
\item{summary.table}{summary table of all events in a particular subtype}

\item{clus.event.mats}{list of submatrices of events for this cluster, indexed by type}

\item{etype}{event type}

\item{full.coverage.df}{merge of all coverage.dfs from genomic saturation result}

\item{cytoband.collapse}{whether or not events on cytobands were merged for the saturation analysis}
}
\value{
matrix for oncoprint plotting
}
\description{
Function to oncoprint submat to plot
Helper function to deal with getting best event for amps/dels
}
\keyword{internal}
