% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{survfit}
\alias{survfit}
\title{Survfit}
\usage{
survfit(full.df, progression.free)
}
\arguments{
\item{full.df}{dataframe of survival data with clusters added}

\item{progression.free}{whether or not to do overall survival or progression free}
}
\value{
dataframe with survObj attached
}
\description{
Helper function to attach survObj to clinical data
}
\keyword{internal}
