% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{survplot.best.v.worst}
\alias{survplot.best.v.worst}
\title{get the best and worst clustering solutions, just plot those}
\usage{
survplot.best.v.worst(clinical.tibble, clustering, output, title.print)
}
\arguments{
\item{clinical.tibble}{Clinical data from TCGA GDAC / Firehose in tibble format}

\item{clustering}{Clustering solution with numeric values as cluster ID, named by samples}

\item{output}{Output .png file for Kaplan-Meyer plot}

\item{title.print}{Plot title}
}
\description{
get the best and worst clustering solutions, just plot those
}
