% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{survplot_best_v_worst}
\alias{survplot_best_v_worst}
\title{Get the best and worst clustering solutions, plot the Kaplan-Meyer of those populations}
\usage{
survplot_best_v_worst(clinical.tibble, clustering, output, title.print)
}
\arguments{
\item{clinical.tibble}{Clinical data from TCGA GDAC / Firehose in tibble format}

\item{clustering}{Clustering solution with numeric values as cluster ID, named by samples}

\item{output}{Output .png file for Kaplan-Meyer plot}

\item{title.print}{Plot title}
}
\value{
Nothing, writes the plot to the supplied .png output file
}
\description{
Get the best and worst clustering solutions, plot the Kaplan-Meyer of those populations
}
