% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble.survfit.progression_free}
\alias{tibble.survfit.progression_free}
\title{progression free survival}
\usage{
tibble.survfit.progression_free(data)
}
\arguments{
\item{data}{Clinical survival data in tibble format, processed from TCGA GDAC / Firehose}
}
\description{
progression free survival
}
