% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble.survfit}
\alias{tibble.survfit}
\title{Add a survival object to the tibble dataset 'survObj' and return}
\usage{
tibble.survfit(data)
}
\arguments{
\item{data}{Clinical survival data in tibble format, processed from TCGA GDAC / Firehose}
}
\value{
The original tibble with a 'survObj' added to the data frame

A copy of the input tibble with a '$survObj' attached
}
\description{
Add a survival object to the tibble dataset 'survObj' and return
}
