% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble.survfit_select}
\alias{tibble.survfit_select}
\title{Get survival based on clustering}
\usage{
tibble.survfit_select(clustering, clinical.tibble,
  progression.free.surv = FALSE)
}
\description{
Get survival based on clustering
}
