% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble_add_clusters}
\alias{tibble_add_clusters}
\title{Add clustering solutions to the tibble dataset 'survObj' and return}
\usage{
tibble_add_clusters(data, clustering)
}
\arguments{
\item{data}{Clinical survival data in tibble format, processed from TCGA GDAC / Firehose}

\item{clustering}{Clustering solution with numeric values as cluster ID, named by samples}
}
\value{
The same tibble with an additional 'clusters' variable added
}
\description{
Add clustering solutions to the tibble dataset 'survObj' and return
}
