% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble_survfit_progression_free}
\alias{tibble_survfit_progression_free}
\title{progression free survival}
\usage{
tibble_survfit_progression_free(data)
}
\arguments{
\item{data}{Clinical survival data in tibble format, processed from TCGA GDAC / Firehose}
}
\value{
A copy of the input tibble with a $survObj attached
}
\description{
progression free survival
}
