% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble_survfit_select}
\alias{tibble_survfit_select}
\title{Get survival based on clustering}
\usage{
tibble_survfit_select(
  clustering,
  clinical.tibble,
  progression.free.surv = FALSE
)
}
\arguments{
\item{clustering}{Clustering solution with numeric values as cluster ID, named by samples}

\item{clinical.tibble}{Clinical data from TCGA GDAC / Firehose in tibble format}

\item{progression.free.surv}{Compute progression-free survival only? (default=FALSE)}
}
\value{
A list containing the survObj, and associated p-values between the best-worst surviving clustrers, as well as overall stratification based on a COX PH model.
}
\description{
Get survival based on clustering
}
