% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{twoStepRankIntegration}
\alias{twoStepRankIntegration}
\title{Two step rank integration}
\usage{
twoStepRankIntegration(interactions.df, na.value)
}
\arguments{
\item{interactions.df}{dataframe of all interactions}

\item{na.value}{value to substitute for NAs if present. Default is NA}
}
\value{
list object with two components: Updated interactions dataframe and a consolidated regulator ranks dataframe
}
\description{
This function does the two step integration of the available genomic information
First Regulator-Event information is integrated across different tests
Then all information for a particular Regulator is integrated to result in a single value
that represents the accumulated information for that Regulator
}
\keyword{internal}
