% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{valid.diggit.interactions}
\alias{valid.diggit.interactions}
\title{Return a set of events 'covered' by specified cMR-event interactions}
\usage{
valid.diggit.interactions(interactions, cnv, gene.loc.mapping,
  selected.tfs)
}
\arguments{
\item{interactions}{: list indexed by amp/mut/del/fus - from cMRs to interacting events}

\item{selected.tfs}{: for each event type list, search within only these cMRS}
}
\value{
a list of events 'covered' by the supplied interactions of type mut/amp/del/fus
}
\description{
Return a set of events 'covered' by specified cMR-event interactions
}
