% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{viper.getSigTFS}
\alias{viper.getSigTFS}
\title{Calculate p-values from pseudo zscores / VIPER aREA scores, threshold}
\usage{
viper.getSigTFS(zscores, fdr.thresh = 0.05)
}
\arguments{
\item{zscores}{Vector of normally distributed z-scores representing protein activities.}

\item{fdr.thresh}{Threshold for false discovery rate, default is 0.05}
}
\description{
Calculate p-values from pseudo zscores / VIPER aREA scores, threshold
}
