% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.interactions.r
\name{viperGetSigTFS}
\alias{viperGetSigTFS}
\title{Calculate p-values from pseudo zscores / VIPER aREA scores, threshold}
\usage{
viperGetSigTFS(zscores, fdr.thresh = 0.05)
}
\arguments{
\item{zscores}{Vector of normally distributed z-scores representing protein activities.}

\item{fdr.thresh}{Threshold for false discovery rate, default is 0.05}
}
\value{
Get the names of proteins with significant z-scores, after multi-hypothesis correction
}
\description{
Calculate p-values from pseudo zscores / VIPER aREA scores, threshold
}
