% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.interactions.r
\name{viperGetTFScores}
\alias{viperGetTFScores}
\title{Function to normalize TF scores}
\usage{
viperGetTFScores(vipermat, fdr.thresh = 0.05)
}
\arguments{
\item{vipermat}{- matrix of VIPER scores with columns as samples, rows as protein names}

\item{fdr.thresh}{- BH-FDR threshold (default 0.05 FDR rate)}
}
\value{
A vector of normalized z-scores, named by TF id
}
\description{
Function to normalize TF scores
}
\examples{
library(moma.gbmexample)
data("gbm.example")
viperGetTFScores(gbm.example$vipermat)
}
