% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{checkGeneMap}
\alias{checkGeneMap}
\title{Check Gene Map}
\usage{
checkGeneMap(gene.loc.mapping)
}
\arguments{
\item{gene.loc.mapping}{dataframe with gene names, entrez ids and cytoband locations}
}
\description{
Check Gene Map
}
\keyword{internal}
