% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.saturation.plots.r
\name{genomicPlotSmall}
\alias{genomicPlotSmall}
\title{Make small genomic plot}
\usage{
genomicPlotSmall(input.df, fraction = 0.85, tissue.cluster = NULL)
}
\arguments{
\item{input.df}{: tissue.coverage.df with mean, k, fraction and unique events.}

\item{fraction}{: what fraction coverage to use for genomic curve threshold}

\item{tissue.cluster}{: which cluster subsample to look at}
}
\value{
output .png
}
\description{
Make small genomic plot
}
\keyword{internal}
