% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma-package.R
\docType{package}
\name{MoMA}
\alias{MoMA}
\alias{MoMA-package}
\title{MoMA: Modern Multivariate Analysis in R}
\description{
TODO
}
\details{
See the package vignettes for details of the algorithm, as well as
comparisons to existing methods.
}
