% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.r
\name{pathway.diggit.intersect}
\alias{pathway.diggit.intersect}
\title{preppi: 3 columns, partner A, B and the p-value of the interaction}
\usage{
pathway.diggit.intersect(diggit.int, pathway, pos.nes.only = TRUE)
}
\arguments{
\item{:}{pathway - a list indexed by TF/MR entrez ID, contains the 
named vector of p-values for interactions}
}
\description{
preppi: 3 columns, partner A, B and the p-value of the interaction
}
