% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{sample.overlap}
\alias{sample.overlap}
\title{The core function to compute which sample-specific alterations overlap with genomic events that are explained 
via DIGGIT.}
\usage{
sample.overlap(momaObj, viper.samples, selected.tfs, interaction.map,
  cnv.threshold = 0.5, mutation.filter = NULL, verbose = FALSE,
  idx.range = NULL)
}
\arguments{
\item{momaObj}{: momaObj}

\item{viper.samples}{: samples within the same cluster}

\item{selected.tfs}{: tfs being analyzed}

\item{interaction.map}{: list object of events 'covered' by the supplied interactions of type mut/amp/del/fus}

\item{mutation.filter}{: a vector of whitelisted mutation events, in entrez gene IDs}

\item{idx.range}{: number of tfs to check for genomic saturation calculation, default is 1253}
}
\description{
The core function to compute which sample-specific alterations overlap with genomic events that are explained 
via DIGGIT.
}
