% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.saturation.r
\name{sample.overlap}
\alias{sample.overlap}
\title{The core function to compute which sample-specific alterations overlap with genomic events that are explained 
via DIGGIT.}
\usage{
sample.overlap(momaObj, viper.samples, selected.tfs, interaction.map,
  cnv.threshold = 0.5, mutation.filter = NULL, verbose = FALSE,
  idx.range = NULL)
}
\arguments{
\item{mutation.filter}{: a vector of whitelisted mutation events, in entrez gene IDs}

\item{:}{diggit.interactions - list indexed by viper protein, includes names of all 
significant interactions}

\item{:}{explain.with -- a set of admissable viper regulators who's diggit interactions may be used to explain
significant interactions}

\item{amp.I}{del.I : map MR names in entrez to genomic locations}

\item{filter.priors}{: use only CHASM and Gistic validated events? Default TRUE
mutation driver prediction based on the nucleotide change}
}
\description{
The core function to compute which sample-specific alterations overlap with genomic events that are explained 
via DIGGIT.
}
