% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{samplename.filter}
\alias{samplename.filter}
\title{Retain TCGA sample ids without the final letter designation ('A/B/C')}
\usage{
samplename.filter(mat)
}
\arguments{
\item{mat}{Matrix of expression or protein activity scores. Columns are sample names, rows are genes}
}
\value{
An identical matrix with new (shorter) column names
}
\description{
Retain TCGA sample ids without the final letter designation ('A/B/C')
}
