% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggit.r
\name{sig.interactors.DIGGIT}
\alias{sig.interactors.DIGGIT}
\title{Filter interactions from NES (DIGGIT) scores and corresponding background-corrected scores. Use this version in the Bayes model to rank TFs}
\usage{
sig.interactors.DIGGIT(corrected.scores, nes.scores, cindy,
  p.thresh = 0.05, cindy.only = TRUE)
}
\arguments{
\item{corrected.scores}{: a list indexed by the genomic event/gene with corresponding pvals and qvals for
each TF}

\item{nes.scores}{: matrix with tfs as columns, rows are genomic events}
}
\value{
a list (indexed by VIPER protein) of significant genomic interactions 
and associated pvals over the background (null TF) model, and NES scores
}
\description{
Filter interactions from NES (DIGGIT) scores and corresponding background-corrected scores. Use this version in the Bayes model to rank TFs
}
