% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{stouffer.integrate.diggit}
\alias{stouffer.integrate.diggit}
\title{Use Stouffer's method to combine z-scores of DIGGIT interactions for each cMR protein.}
\usage{
stouffer.integrate.diggit(interactions, from.p = FALSE, pos.nes.only = TRUE)
}
\arguments{
\item{interactions}{A list indexed by TF, includes z-scores or p-values for each interacting event}

\item{from.p}{Integrate p-values or z-scores (default z-scores; from.p = FALSE)}

\item{pos.nes.only}{Use only positive NES scores to rank proteins (default TRUE)}
}
\value{
A list indexed by TF, a stouffer integrated z-score
}
\description{
This function combines only positively associated DIGGIT scores by default to create a culmulative DIGGIT score for each cMR.
}
