% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.r
\name{stouffer.integrate}
\alias{stouffer.integrate}
\title{dispatch method for either CNV location corrected or SNV}
\usage{
stouffer.integrate(interactions, cytoband.map = NULL)
}
\arguments{
\item{interactions}{List of MR - Genomic Event interactions, inferred by DIGGIT}

\item{cytoband.map}{Data.frame mapping Entrez.IDs to cytoband locations}
}
\value{
Z-scores for each MR
}
\description{
dispatch method for either CNV location corrected or SNV
}
