% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{survplot.best.v.worst}
\alias{survplot.best.v.worst}
\title{get the best and worst clustering solutions, just plot those}
\usage{
survplot.best.v.worst(clinical.tibble, clustering, output, title.print)
}
\description{
get the best and worst clustering solutions, just plot those
}
