% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble.add_clusters}
\alias{tibble.add_clusters}
\title{Add clustering solutions to the tibble dataset 'survObj' and return}
\usage{
tibble.add_clusters(data, clustering)
}
\description{
Add clustering solutions to the tibble dataset 'survObj' and return
}
