% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.surv.r
\name{tibble.survfit}
\alias{tibble.survfit}
\title{Add a survival object to the tibble dataset 'survObj' and return}
\usage{
tibble.survfit(data)
}
\description{
Add a survival object to the tibble dataset 'survObj' and return
}
